feat: añade test timeFormat y elimina dependencias en desuso
