+++ package.json
@@ -6,11 +6,12 @@
         "@testing-library/jest-dom": "^5.11.4",
         "@testing-library/react": "^11.1.0",
         "@testing-library/user-event": "^12.1.10",
-        "reactstrap": "^9.0.1",
+        "babel-plugin-transform-react-constant-elements": "^6.23.0",
         "bootstrap": "^5.1.3",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
         "react-scripts": "4.0.2",
+        "reactstrap": "^9.0.1",
         "web-vitals": "^1.0.1"
     },
     "scripts": {
@@ -36,15 +37,5 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
-    },
-    "babel": {
-        "env": {
-            "production": {
-                "plugins": [
-                    "transform-react-constant-elements",
-                    "transform-react-inline-elements"
-                ]
-            }
-        }
     }
 }
