+++ src/components/Timer.js
@@ -1,16 +1,5 @@
 import { useState, useEffect } from "react";
-
-const timeFormat = (date) => {
-  if (!date) return "00:00";
-
-  let mm = date.getUTCMinutes();
-  let ss = date.getSeconds();
-
-  mm = mm < 10 ? "0" + mm : mm;
-  ss = ss < 10 ? "0" + ss : ss;
-
-  return `${mm}:${ss}`;
-};
+import { timeFormat } from "../utils/timeFormat";
 
 export default function Timer({ gameOver, handleTimeRecord }) {
   const [init] = useState(+new Date());
